requirements.txt
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
beautifulsoup4==4.12.2
lxml
pandas
numpy
psycopg2-binary
dateparser==1.2.0